condensing the application contexts in the UI to a single once since they were identical
