+++ web-client/src/index.prod.js
@@ -1,7 +1,7 @@
 import '@babel/polyfill';
 
 import app from './app';
-import applicationContext from './applicationContexts/prod';
+import applicationContext from './applicationContext';
 
 /**
  * Initializes the app with prod environment context
